Migrate shifted_gates method from QubitExcitation to GeneralizedRotation.
